Finished Project
I completed the color scheme of the website, fixed the quotes (to an extent) at the end of the end blockquote element. I removed the section#rightcol background color and replaced it with a body background color. I centered the text in the section#rightcol. I centered the text in the second aside element in the section#leftcol. I also made it so that when the mouse is above the links in the section#rightcol, they are enlarged.